user_database/requirements.txt
psycopg2-binary
flask-sqlalchemy
Flask-Migrate
flask-script
flask_cors